Refactor dsl modules organization and scaladoc
