+++ core/src/main/scala/dev/atedeg/ecscala/dsl/ECScalaDSL.scala
@@ -1,8 +1,9 @@
 package dev.atedeg.ecscala.dsl
 
+import dev.atedeg.ecscala.dsl.Words.{ ClearWord, EntityWord }
+import dev.atedeg.ecscala.dsl.Syntax
 import dev.atedeg.ecscala.util.types.{ CListTag, ComponentTag }
 import dev.atedeg.ecscala.{ CList, CNil, Component, Deletable, DeltaTime, Entity, System, View, World }
-import dev.atedeg.ecscala.dsl.Words.*
 
 /**
  * This trait provides a domain specific language (DSL) for expressing the ECScala framework operations using an
@@ -73,7 +74,7 @@ import dev.atedeg.ecscala.dsl.Words.*
  *   clearAll from world
  * }}}
  */
-trait ECScalaDSL extends ExtensionMethods with Conversions with FromSyntax {
+trait ECScalaDSL extends ExtensionMethods with Conversions with Syntax {
 
   /**
    * Keyword that enables the use of the word "entity" in the dsl.
@@ -110,101 +111,4 @@ trait ECScalaDSL extends ExtensionMethods with Conversions with FromSyntax {
    * Keyword that enables the use of the word "clearAll" in the dsl.
    */
   def clearAll: From[World, Unit] = ClearAllFromWorld(ClearWord())
-}
-
-private[dsl] trait FromSyntax {
-
-  /**
-   * This trait enables the use of the word "from" in the dsl
-   */
-  trait From[A, B] {
-    def from(elem: A): B
-  }
-
-  /**
-   * This case class enables the following syntax:
-   * {{{
-   *   remove (entity1, entity2) from world
-   * }}}
-   */
-  class EntitiesFromWorld(entities: Seq[Entity]) extends From[World, Unit] {
-    override def from(world: World): Unit = entities foreach { world.removeEntity(_) }
-  }
-
-  /**
-   * This case class enables the following syntax:
-   * {{{
-   *   remove (system1) from world
-   * }}}
-   */
-  class SystemFromWorld[L <: CList: CListTag](system: System[L]) extends From[World, Unit] {
-    override def from(world: World): Unit = world.removeSystem(system)
-  }
-
-  /**
-   * This case class enables the following syntax:
-   * {{{
-   *   clearAll from world
-   * }}}
-   */
-  class ClearAllFromWorld(clearWord: ClearWord) extends From[World, Unit] {
-    override def from(world: World): Unit = world.clear()
-  }
-
-  /**
-   * This case class enables the following syntax:
-   *
-   * {{{
-   *   remove (myComponent) from entity1
-   * }}}
-   */
-  class FromEntity[L <: CList](componentList: L)(using clt: CListTag[L]) extends From[Entity, Unit] {
-
-    override def from(entity: Entity): Unit =
-      componentList zip clt.tags.asInstanceOf[Seq[ComponentTag[Component]]] foreach {
-        entity.removeComponent(_)(using _)
-      }
-  }
-
-  /**
-   * This case class enables the following syntax:
-   *
-   * {{{
-   *   * getView[MyComponent1 &: MyComponent2 &: CNil] from world
-   *   * getView[MyComponent1 &: MyComponent2 &: CNil].exluding[MyComponent3 &: CNil] from world
-   * }}}
-   */
-  class ViewFromWorld[A <: CList](using cltA: CListTag[A]) extends From[World, View[A]] {
-    override def from(world: World): View[A] = world.getView(using cltA)
-
-    def excluding[B <: CList](using cltB: CListTag[B]): ExcludingViewFromWorld[A, B] = ExcludingViewFromWorld(using
-      cltA,
-    )(using cltB)
-  }
-
-  class ExcludingViewFromWorld[A <: CList, B <: CList](using cltA: CListTag[A])(using cltB: CListTag[B])
-      extends From[World, View[A]] {
-    def from(world: World) = world.getView(using cltA, cltB)
-  }
-}
-
-object Words {
-
-  /**
-   * This case class enables the following syntax:
-   *
-   * {{{
-   * world hasAn entity
-   * }}}
-   */
-  case class EntityWord()
-
-  /**
-   * This case class enables the following syntax:
-   *
-   * {{{
-   * clearAll from world
-   * }}}
-   */
-  private[dsl] case class ClearWord()
 }
